Move attachment endpoints into new namespace
Also, format arg lists and remove unused usings.
